# perf(http): split up `op_http_get_request_method_and_url`