--- a/ext/http/00_serve.ts
+++ b/ext/http/00_serve.ts
@@ -12,13 +12,14 @@ import {
   op_http_close,
   op_http_close_after_finish,
   op_http_get_request_headers,
-  op_http_get_request_method_and_url,
+  op_http_get_request_method_authority_and_path,
   op_http_read_request_body,
   op_http_serve,
   op_http_serve_on,
   op_http_set_promise_complete,
   op_http_set_response_body_bytes,
   op_http_set_response_body_resource,
+  op_http_get_request_peer_address_and_port,
   op_http_set_response_body_text,
   op_http_set_response_header,
   op_http_set_response_headers,
@@ -134,7 +135,8 @@ function addTrailers(resp, headerList) {
 class InnerRequest {
   #external;
   #context;
-  #methodAndUri;
+  #methodAuthorityAndPath;
+  #peerAddressAndPort;
   #streamRid;
   #body;
   #upgraded;
@@ -255,16 +257,14 @@ class InnerRequest {
       return this.#urlValue;
     }
 
-    if (this.#methodAndUri === undefined) {
+    if (this.#methodAuthorityAndPath === undefined) {
       if (this.#external === null) {
         throw new TypeError("Request closed");
       }
-      // TODO(mmastrac): This is quite slow as we're serializing a large number of values. We may want to consider
-      // splitting this up into multiple ops.
-      this.#methodAndUri = op_http_get_request_method_and_url(this.#external);
+      this.#methodAuthorityAndPath = op_http_get_request_method_authority_and_path(this.#external);
     }
 
-    const path = this.#methodAndUri[2];
+    const path = this.#methodAuthorityAndPath[2];
 
     // * is valid for OPTIONS
     if (path === "*") {
@@ -273,15 +273,15 @@ class InnerRequest {
 
     // If the path is empty, return the authority (valid for CONNECT)
     if (path == "") {
-      return this.#urlValue = this.#methodAndUri[1];
+      return this.#urlValue = this.#methodAuthorityAndPath[1];
     }
 
     // CONNECT requires an authority
-    if (this.#methodAndUri[0] == "CONNECT") {
-      return this.#urlValue = this.#methodAndUri[1];
+    if (this.#methodAuthorityAndPath[0] === "CONNECT") {
+      return this.#urlValue = this.#methodAuthorityAndPath[1];
     }
 
-    const hostname = this.#methodAndUri[1];
+    const hostname = this.#methodAuthorityAndPath[1];
     if (hostname) {
       // Construct a URL from the scheme, the hostname, and the path
       return this.#urlValue = this.#context.scheme + hostname + path;
@@ -313,27 +313,27 @@ class InnerRequest {
         path: this.#context.listener.addr.path,
       };
     }
-    if (this.#methodAndUri === undefined) {
+    if (this.#peerAddressAndPort === undefined) {
       if (this.#external === null) {
         throw new TypeError("Request closed");
       }
-      this.#methodAndUri = op_http_get_request_method_and_url(this.#external);
+      this.#peerAddressAndPort = op_http_get_request_peer_address_and_port(this.#external);
     }
     return {
       transport: "tcp",
-      hostname: this.#methodAndUri[3],
-      port: this.#methodAndUri[4],
+      hostname: this.#peerAddressAndPort[0],
+      port: this.#peerAddressAndPort[1],
     };
   }
 
   get method() {
-    if (this.#methodAndUri === undefined) {
+    if (this.#methodAuthorityAndPath === undefined) {
       if (this.#external === null) {
         throw new TypeError("Request closed");
       }
-      this.#methodAndUri = op_http_get_request_method_and_url(this.#external);
+      this.#methodAuthorityAndPath = op_http_get_request_method_authority_and_path(this.#external);
     }
-    return this.#methodAndUri[0];
+    return this.#methodAuthorityAndPath[0];
   }
 
   get body() {