```ts
Deno.serve({ port: 8085 }, request => {
  return new Response(request.url);
});
```

Before:

```
% wrk -d60s http://localhost:8085
Running 1m test @ http://localhost:8085
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    54.72us   14.00us   2.23ms   86.98%
    Req/Sec    83.48k     2.95k   88.37k    90.52%
  9984078 requests in 1.00m, 1.50GB read
Requests/sec: 166124.33
Transfer/sec:     25.51MB
```

After:

```
% wrk -d60s http://localhost:8085
Running 1m test @ http://localhost:8085
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    54.33us   12.95us   2.00ms   84.53%
    Req/Sec    84.00k     3.10k   90.98k    91.26%
  10046685 requests in 1.00m, 1.51GB read
Requests/sec: 167166.17
Transfer/sec:     25.67MB
```