@dsherret 2024-10-15
I'm going to retry these benchmarks on my mac because I think my linux system is unstable.